Remove Searching interface and class implementation and move search by param method to Contact service
