+++ crm-common/service-maper/src/main/java/com/example/dto/SaveContactDto.java
@@ -7,19 +7,9 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class SaveContactDto {
-
-    private String firstName;
-
-    private String lastName;
-
-    private String jobTitle;
-
-    private String company;
+public class SaveContactDto extends ContactDto {
 
     private String country;
-
-    private String email;
 
     private String mobilePhone;
 }
