--- crm-common/service/src/main/java/com/example/service/ContactSearchingService.java
@@ -1,10 +1,0 @@
-package com.example.service;
-
-import com.example.Contact;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-public interface ContactSearchingService {
-
-    Page<Contact> findAllByParam(final String filteredBy, final String query, final Pageable pageable);
-}
